api/pyproject.toml
[tool.poetry]
name = "zen-visualization-api"
version = "0.1.0"
description = ""
authors = ["Stian Nyblom <stiannyb@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
pandas = "^2.2.1"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
influxdb-client = "^1.42.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
ruff = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
